получение списка проектов и задач
