--- data.php
@@ -1,41 +1,0 @@
-<?php
-
-/* $tasks = [
-    [
-        'name' => 'Собеседование в IT компании',
-        'date' => '01.03.2021',
-        'project' => 'Работа',
-        'status' => false,
-    ],
-    [
-        'name' => 'Выполнить тестовое задание',
-        'date' => '04.02.2021',
-        'project' => 'Работа',
-        'status' => false
-    ],
-    [
-        'name' => 'Сделать задание первого раздела',
-        'date' => '21.12.2020',
-        'project' => 'Учёба',
-        'status' => true
-    ],
-    [
-        'name' => 'Встреча с другом',
-        'date' => '22.02.2021',
-        'project' => 'Входящие',
-        'status' => false
-    ],
-    [
-        'name' => 'Купить корм для кота',
-        'date' => '31.01.2021',
-        'project' => 'Домашние дела',
-        'status' => false
-    ],
-    [
-        'name' => 'Заказать пиццу',
-        'date' => null,
-        'project' => 'Домашние дела',
-        'status' => false
-    ],
-
-]; */
